+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+-   Adds a new command, `renpyWarp.jumpToLabel`, which allows you to jump to a
+    label in your Ren'Py project
+
 ## 1.16.0 - 2024-08-18
 
 -   Now focuses window and warp in parallel, which should make it more